backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyyaml==6.0.1
requests==2.31.0
aiohttp==3.13.3
psycopg2-binary==2.9.9
python-dotenv==1.2.1